okay everything works in a class structure
