+++ index.html
@@ -14,7 +14,6 @@
     <div class="instructions-container">
       <h1 id="instructions"></h1>
     </div>
-    <button type="button" id="ai-button" state="off">AI Player: off</button>
   </body>
 
   <!-- Scripts -->
@@ -22,8 +21,8 @@
   <script src="./lib/two.min.js" charset="utf-8"></script>
   <script src="./lib/jquery-2.2.0.min.js" charset="utf-8"></script>
     <!-- actual stuff -->
-  <script src="./dist/num.js" charset="utf-8"></script>
-  <script src="./dist/keybindings.js" charset="utf-8"></script>
-  <script src="./dist/main.js" charset="utf-8"></script>
+  <!-- <script src="./dist/num.js" charset="utf-8"></script> -->
+  <!-- <script src="./dist/keybindings.js" charset="utf-8"></script> -->
+  <script src="./build/bundle.js" charset="utf-8"></script>
 
 </html>
