+++ js/main.js
@@ -2,158 +2,126 @@
 //jshint -W097
 'use strict';
 
-let isMobile = typeof window.orientation !== 'undefined'
-window.GAMERUNNING = true
-window.AIPLAYING = false
-window.AIWORKER = null
+// require('../lib/jquery-2.2.0.min.js')
+// require('../lib/two.min.js')
+require('./keybindings.js')
+// require('./util.js')
 
-// Make an instance of two and place it on the page.
-let elem = document.getElementById('draw-shapes')
-let params = { width: window.innerWidth, height: window.innerHeight }
-let two = new Two(params).appendTo(elem)
+import * as d3 from "d3"
+import * as utils from './util.js';
 
-let centerRadius = two.height / 8
-let centerX = two.width / 2
-let centerY = two.height / 2
-let thickness = two.height / 20
-console.log(thickness);
+class Game {
+  constructor() {
+    this.isMobile = typeof window.orientation !== 'undefined'
+    this.running = true
 
-let background = two.makeRectangle(centerX, centerY, two.width, two.height)
-let centerCircle = two.makeCircle(centerX, centerY,  centerRadius)
-let smallCircle = two.makeEllipse(
-  centerX - centerRadius, centerY,
-  thickness * 0.4, thickness * 0.4
-)
-let tick = two.makeRoundedRectangle(
-  centerX - centerRadius, centerY,
-  thickness / 2, thickness * 1.9, 3
-)
+    // Make an instance of two and place it on the page.
+    this.elem = document.getElementById('draw-shapes')
+    this.params = { width: window.innerWidth, height: window.innerHeight }
+    this.two = new Two(this.params).appendTo(this.elem)
 
-background.fill = 'rgb(109, 140, 149)'
-background.noStroke()
+    this.centerRadius = this.two.height / 8
+    this.centerX = this.two.width / 2
+    this.centerY = this.two.height / 2
+    this.thickness = this.two.height / 20
 
-centerCircle.noFill()
-centerCircle.stroke = 'rgb(54, 60, 77)'
-centerCircle.linewidth = thickness
+    this.background = this.two.makeRectangle(this.centerX, this.centerY, this.two.width, this.two.height)
+    this.centerCircle = this.two.makeCircle(this.centerX, this.centerY,  this.centerRadius)
+    this.smallCircle = this.two.makeEllipse(
+      this.centerX - this.centerRadius, this.centerY,
+      this.thickness * 0.4, this.thickness * 0.4
+    )
+    this.tick = this.two.makeRoundedRectangle(
+      this.centerX - this.centerRadius, this.centerY,
+      this.thickness / 2, this.thickness * 1.9, 3
+    )
 
-smallCircle.fill = 'rgb(228, 142, 13)'
-smallCircle.noStroke()
+    this.background.fill = 'rgb(109, 140, 149)'
+    this.background.noStroke()
 
-tick.rotation = Math.PI / 2
-tick.noStroke()
-tick.fill = 'rgb(81, 191, 222)'
+    this.centerCircle.noFill()
+    this.centerCircle.stroke = 'rgb(54, 60, 77)'
+    this.centerCircle.linewidth = this.thickness
 
-let highScore = parseFloat(getCookie('highScore2')) || 0
-let tickAngle = 0
-let points = 0
-let randAngle = 2 * Math.PI * Math.random()
-let direction = 1
+    this.smallCircle.fill = 'rgb(228, 142, 13)'
+    this.smallCircle.noStroke()
 
-updateScore()
+    this.tick.rotation = Math.PI / 2
+    this.tick.noStroke()
+    this.tick.fill = 'rgb(81, 191, 222)'
 
-two.bind('update', function(frameCount) {
-  if (GAMERUNNING) {
-    updateGame()
-    updateAIWorker()
+    this.highScore = parseFloat(this.getCookie('highScore2')) || 0
+    this.tickAngle = 0
+    this.points = 0
+    this.randAngle = 2 * Math.PI * Math.random()
+    this.direction = 1
+
+    this.updateScore()
+
+    this.two.bind('update', (frameCount) => {
+      if (this.running) {
+        this.updateGame()
+      }
+    }).play()
+
+    let ten = (10 * (Math.PI / 180)) // ten degrees in radians
+
+    $(window).on('key-esc', () => this.running = !this.running )
+
+    $(window).on('key-space', () => {
+      if (this.tick.rotation > this.smallCircle.rotation - ten &&
+          this.tick.rotation < this.smallCircle.rotation + ten) {
+        this.direction *= -1
+        this.randAngle += Math.PI * Math.random()
+        this.randAngle %= 2 * Math.PI
+        this.points += 1
+      } else {
+        this.points = 0
+      }
+
+      this.updateScore()
+    })
   }
-}).play()
 
-let ten = (10 * (Math.PI / 180)) // ten degrees in radians
+  updateGame() {
+    let tickSpeed = this.points * 0.001
+    this.tickAngle += (0.05 + tickSpeed) * this.direction; this.tickAngle %= 2 * Math.PI
 
-$(window).on('key-r', () => { document.getElementById('ai-button').onclick() })
+    this.tick.translation.x = this.centerX + (this.centerRadius * Math.cos(this.tickAngle))
+    this.tick.translation.y = this.centerY + (this.centerRadius * Math.sin(this.tickAngle))
+    this.tick.rotation = (Math.PI / 2) + Math.atan2(
+      Math.sin(this.tickAngle),
+      Math.cos(this.tickAngle)
+    )
 
-$(window).on('key-esc', () => { GAMERUNNING = !GAMERUNNING })
-
-$(window).on('key-space', () => {
-  if (tick.rotation > smallCircle.rotation - ten &&
-      tick.rotation < smallCircle.rotation + ten) {
-    direction *= -1
-    randAngle += Math.PI * Math.random()
-    randAngle %= 2 * Math.PI
-    points += 1
-  } else {
-    points = 0
+    this.smallCircle.translation.x = this.centerX + (this.centerRadius * Math.cos(this.randAngle))
+    this.smallCircle.translation.y = this.centerY + (this.centerRadius * Math.sin(this.randAngle))
+    this.smallCircle.rotation = (Math.PI / 2) + Math.atan2(
+      Math.sin(this.randAngle),
+      Math.cos(this.randAngle)
+    )
   }
 
-  updateScore()
-})
-
-let prevMessageTime = 0
-
-function onAIMessage(event) {
-  // console.log('received message from AI worker')
-  // console.log(event.timeStamp - prevMessageTime > 350)
-  if (event.timeStamp - prevMessageTime > 350) {
-    // console.log(event.data);
-    $(window).trigger(event.data)
-    prevMessageTime = event.timeStamp
+  updateScore() {
+    this.highScore = (this.highScore > this.points ? this.highScore : this.points)
+    document.getElementById('high-score').innerHTML = 'High Score: ' + this.highScore
+    document.getElementById('score').innerHTML = this.points
+    this.setCookie('highScore2', this.highScore, 9999999999999)
+    let instructions = document.getElementById('instructions')
+    if (this.points === 0)
+      instructions.innerHTML = `${ this.isMobile ? 'tap' : 'press the spacebar'} to begin`
+    else
+      instructions.innerHTML = ''
   }
-}
 
-function updateAIWorker() {
-  if (AIPLAYING)
-    AIWORKER.postMessage([tick.rotation, smallCircle.rotation])
-}
-
-function updateGame() {
-  let tickSpeed = points * 0.001
-  tickAngle += (0.05 + tickSpeed) * direction; tickAngle %= 2 * Math.PI
-
-  tick.translation.x = centerX + (centerRadius * Math.cos(tickAngle))
-  tick.translation.y = centerY + (centerRadius * Math.sin(tickAngle))
-  tick.rotation = (Math.PI / 2) + Math.atan2(
-    Math.sin(tickAngle),
-    Math.cos(tickAngle)
-  )
-
-  smallCircle.translation.x = centerX + (centerRadius * Math.cos(randAngle))
-  smallCircle.translation.y = centerY + (centerRadius * Math.sin(randAngle))
-  smallCircle.rotation = (Math.PI / 2) + Math.atan2(
-    Math.sin(randAngle),
-    Math.cos(randAngle)
-  )
-}
-
-function updateScore() {
-  highScore = (highScore > points ? highScore : points)
-  document.getElementById('high-score').innerHTML = 'High Score: ' + highScore
-  document.getElementById('score').innerHTML = points
-  setCookie('highScore2', highScore, 999999)
-  let instructions = document.getElementById('instructions')
-  if (points === 0){
-    // console.log(isMobile ? 'tap' : 'press the spacebar');
-    instructions.innerHTML = `${ isMobile ? 'tap' : 'press the spacebar'} to begin`
-  }
-  else instructions.innerHTML = ''
-}
-
-document.getElementById('ai-button').onclick = function() {
-  AIPLAYING = !AIPLAYING
-
-  let button = document.getElementById('ai-button')
-  button.innerHTML = 'AI Player: ' + (AIPLAYING ? 'on' : 'off')
-  button.setAttribute('state', (AIPLAYING ? 'on' : 'off'))
-  button.blur()
-
-  if (AIPLAYING) {
-    AIWORKER = new Worker("./js/ai-player.js");
-    AIWORKER.onmessage = onAIMessage
-
-  } else {
-    AIWORKER.terminate()
-    AIWORKER = null
-  }
-}
-
-// http://www.w3schools.com/js/js_cookies.asp
-function setCookie(cname, cvalue, exdays) {
+  setCookie(cname, cvalue, exdays) {
     var d = new Date();
     d.setTime(d.getTime() + (exdays*24*60*60*1000));
     var expires = "expires="+d.toUTCString();
     document.cookie = cname + "=" + cvalue + "; " + expires;
-}
+  }
 
-function getCookie(cname) {
+  getCookie(cname) {
     var name = cname + "=";
     var ca = document.cookie.split(';');
     for(var i = 0; i<ca.length; i++) {
@@ -162,4 +130,7 @@ function getCookie(cname) {
         if (c.indexOf(name) === 0) return c.substring(name.length,c.length);
     }
     return "";
+  }
 }
+
+new Game()
