+++ webpack.config.js
@@ -1,0 +1,20 @@
+module.exports = {
+  entry: './js/main.js',
+  output: {
+    path: __dirname + '/build',
+    publicPath: '/build/',
+    filename: 'bundle.js',
+  },
+  module: {
+    loaders: [
+      {
+        test: /\.js$/,
+        exclude: /(node_modules|bower_components)/,
+        loader: 'babel',
+        query: {
+          presets: ['es2015']
+        }
+      }
+    ]
+  }
+}
